requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
pytube==15.0.0
openai-whisper==20231117
ffmpeg-python==0.2.0 